requirements.txt
# Saturday MK1 — Production Dependencies
# =========================================
# Core (required for Vercel deployment)
pyyaml>=6.0.1
fastapi>=0.115.0
pydantic>=2.9.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
